requirements.txt
boto3==1.34.144
environs==9.5.0
fastapi==0.111.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pytest==7.4.0
PyJWT==2.8.0
requests==2.31.0
uvicorn==0.23.2
